requirements.txt
aiosqlite==0.16.1
arrow==0.17.0
atomicwrites==1.4.0
attrs==21.2.0
binaryornot==0.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.2
fastapi==0.65.2
h11==0.11.0
idna==2.10
iniconfig==1.1.1
iso8601==0.1.14
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
poyo==0.5.0
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.4
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
requests==2.24.0
six==1.15.0
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
urllib3==1.25.10
uvicorn==0.12.1
wincertstore==0.2
